# Optimisations On 10M Calls