@martindevans 2023-12-18
I initially had this code guarding access to the context:

```csharp
if (handle.IsInvalid)
{
    throw new ObjectDisposedException(typeof(Store).FullName);
}

return new StoreContext(contextHandle);
```

However, this failed in unit tests. Even after `store.Dispose()` is called `IsInvalid` was _false_ but `IsClosed` was _true_.

I'm not sure if this indicates a potential issue in other places `IsInvalid` is used?