@kpreisser 2023-12-20
> However, this failed in unit tests. Even after `store.Dispose()` is called `IsInvalid` was _false_ but `IsClosed` was _true_.
> 
> I'm not sure if this indicates a potential issue in other places `IsInvalid` is used?

AFAIK, `SafeHandle.IsInvalid` determines whether the handle value is invalid according to the handle type it represents (e.g. `SafeHandleZeroOrMinusOneIsInvalid` would report `0` and `-1` as being invalid); it doesn't necessarily report whether the handle has already been released/disposed, which is indicated by [`.IsClosed`](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.interopservices.safehandle.isclosed?view=net-8.0).

If `SafeHandle.IsInvalid` is used in other places where the intention was to check whether the handle has already been released (e.g. to throw an `ObjectDisposedException`), then I would agree this was probably a mistake and should be replaced by calling `.IsClosed`.