@martindevans 2023-12-21
@kpreisser in this case [here's](https://github.com/bytecodealliance/wasmtime-dotnet/blob/main/src/Store.cs#L282) an example of some code that looks suspect to me. This same pattern is used everywhere, in fact the new code I just added is the _only_ placed `IsClosed` is used! Test coverage shows that none of the `IsInvalid` branches are ever tested.