@martindevans 2023-12-21
Opened https://github.com/bytecodealliance/wasmtime-dotnet/pull/289 to resolve the IsInvalid thing.